refactor: move changeCowBreedingPenResident to its own file
