+++ src/game-logic/reducers/index.js
@@ -11,7 +11,6 @@ import {
   nullArray,
 } from '../../utils'
 import {
-  COW_GESTATION_PERIOD_DAYS,
   COW_HUG_BENEFIT,
   MAX_ANIMAL_NAME_LENGTH,
   MAX_DAILY_COW_HUG_BENEFITS,
@@ -84,51 +83,6 @@ export * from './sellCow'
 export * from './removeCowFromInventory'
 export * from './modifyCow'
 export * from './changeCowAutomaticHugState'
-
-/**
- * @param {farmhand.state} state
- * @param {farmhand.cow} cow
- * @param {boolean} doAdd If true, cow will be added to the breeding pen. If
- * false, they will be removed.
- * @returns {farmhand.state}
- */
-export const changeCowBreedingPenResident = (state, cow, doAdd) => {
-  const { cowBreedingPen } = state
-  const { cowId1, cowId2 } = cowBreedingPen
-  const isPenFull = cowId1 !== null && cowId2 !== null
-  const isCowInPen = cowId1 === cow.id || cowId2 === cow.id
-  let newCowBreedingPen = { ...cowBreedingPen }
-
-  if (doAdd) {
-    if (isPenFull || isCowInPen) {
-      return state
-    }
-
-    const cowId = cowId1 === null ? 'cowId1' : 'cowId2'
-    newCowBreedingPen = { ...newCowBreedingPen, [cowId]: cow.id }
-  } else {
-    if (!isCowInPen) {
-      return state
-    }
-
-    if (cowId1 === cow.id) {
-      newCowBreedingPen = {
-        ...newCowBreedingPen,
-        cowId1: newCowBreedingPen.cowId2,
-      }
-    }
-
-    newCowBreedingPen = { ...newCowBreedingPen, cowId2: null }
-  }
-
-  return {
-    ...state,
-    cowBreedingPen: {
-      ...newCowBreedingPen,
-      daysUntilBirth: COW_GESTATION_PERIOD_DAYS,
-    },
-  }
-}
 
 /**
  * @param {farmhand.state} state
