+++ src/game-logic/reducers/index.test.js
@@ -5,7 +5,6 @@ import {
   LOAN_PAYOFF,
 } from '../../templates'
 import {
-  COW_GESTATION_PERIOD_DAYS,
   COW_HUG_BENEFIT,
   MAX_ANIMAL_NAME_LENGTH,
   MAX_LATEST_PEER_MESSAGES,
@@ -42,168 +41,6 @@ describe('selectCow', () => {
   test('updates selectedCowId', () => {
     const { selectedCowId } = fn.selectCow({}, { id: 'abc' })
     expect(selectedCowId).toEqual('abc')
-  })
-})
-
-describe('changeCowBreedingPenResident', () => {
-  describe('doAdd === false', () => {
-    describe('cow is not in breeding pen', () => {
-      test('no-ops', () => {
-        const inputState = {
-          cowBreedingPen: {
-            cowId1: 'cow-a',
-            cowId2: 'cow-b',
-            daysUntilBirth: COW_GESTATION_PERIOD_DAYS,
-          },
-        }
-
-        const state = fn.changeCowBreedingPenResident(
-          inputState,
-          generateCow({ id: 'cow-c' }),
-          false
-        )
-
-        expect(state).toBe(inputState)
-      })
-    })
-
-    describe('cow is in position 1', () => {
-      test('cow is removed', () => {
-        const state = fn.changeCowBreedingPenResident(
-          {
-            cowBreedingPen: {
-              cowId1: 'cow-a',
-              cowId2: 'cow-b',
-              daysUntilBirth: COW_GESTATION_PERIOD_DAYS,
-            },
-          },
-          generateCow({ id: 'cow-a' }),
-          false
-        )
-
-        expect(state).toEqual({
-          cowBreedingPen: {
-            cowId1: 'cow-b',
-            cowId2: null,
-            daysUntilBirth: COW_GESTATION_PERIOD_DAYS,
-          },
-        })
-      })
-    })
-
-    describe('cow is in position 2', () => {
-      test('cow is removed', () => {
-        const state = fn.changeCowBreedingPenResident(
-          {
-            cowBreedingPen: {
-              cowId1: 'cow-a',
-              cowId2: 'cow-b',
-              daysUntilBirth: COW_GESTATION_PERIOD_DAYS,
-            },
-          },
-          generateCow({ id: 'cow-b' }),
-          false
-        )
-
-        expect(state).toEqual({
-          cowBreedingPen: {
-            cowId1: 'cow-a',
-            cowId2: null,
-            daysUntilBirth: COW_GESTATION_PERIOD_DAYS,
-          },
-        })
-      })
-    })
-  })
-
-  describe('doAdd === true', () => {
-    describe('cow is in breeding pen', () => {
-      test('no-ops', () => {
-        const inputState = {
-          cowBreedingPen: {
-            cowId1: 'cow-a',
-            cowId2: 'cow-b',
-            daysUntilBirth: COW_GESTATION_PERIOD_DAYS,
-          },
-        }
-
-        const state = fn.changeCowBreedingPenResident(
-          inputState,
-          generateCow({ id: 'cow-a' }),
-          true
-        )
-
-        expect(state).toBe(inputState)
-      })
-    })
-
-    describe('breeding pen is full', () => {
-      test('no-ops', () => {
-        const inputState = {
-          cowBreedingPen: {
-            cowId1: 'cow-a',
-            cowId2: 'cow-b',
-            daysUntilBirth: COW_GESTATION_PERIOD_DAYS,
-          },
-        }
-
-        const state = fn.changeCowBreedingPenResident(
-          inputState,
-          generateCow({ id: 'cow-c' }),
-          true
-        )
-
-        expect(state).toBe(inputState)
-      })
-    })
-
-    describe('there are no cows in breeding pen', () => {
-      test('cow is added to first slot', () => {
-        const state = fn.changeCowBreedingPenResident(
-          {
-            cowBreedingPen: {
-              cowId1: null,
-              cowId2: null,
-              daysUntilBirth: COW_GESTATION_PERIOD_DAYS,
-            },
-          },
-          generateCow({ id: 'cow-a' }),
-          true
-        )
-
-        expect(state).toEqual({
-          cowBreedingPen: {
-            cowId1: 'cow-a',
-            cowId2: null,
-            daysUntilBirth: COW_GESTATION_PERIOD_DAYS,
-          },
-        })
-      })
-    })
-
-    describe('there is one cow in breeding pen', () => {
-      test('cow is added to second slot', () => {
-        const state = fn.changeCowBreedingPenResident(
-          {
-            cowBreedingPen: {
-              cowId1: 'cow-a',
-              cowId2: null,
-              daysUntilBirth: COW_GESTATION_PERIOD_DAYS,
-            },
-          },
-          generateCow({ id: 'cow-b' }),
-          true
-        )
-
-        expect(state).toEqual({
-          cowBreedingPen: {
-            cowId1: 'cow-a',
-            cowId2: 'cow-b',
-            daysUntilBirth: COW_GESTATION_PERIOD_DAYS,
-          },
-        })
-      })
-    })
   })
 })
